Build the Go agent as part of the SDK app build
The agent was the last piece still built by hand. `make` in this directory now
produces both binaries and stages them into $(RK_APP_OUTPUT)/bin, so a plain
./build.sh puts motionsense-agent in /oem/usr/bin alongside MotionSense.

It cross-compiles with GOARM=7 and CGO_ENABLED=0 — modernc.org/sqlite is a
pure-Go implementation, so nothing has to link against uclibc and the binary
is static. -trimpath -ldflags="-s -w" takes it from 15.7MB to 11MB.

Two ordering constraints are encoded in the Makefile rather than left to luck:

- go-agent depends on c-daemon purely for sequencing. Makefile.param sets
  MAKEFLAGS += -j, so without it the two targets race and the strip step can
  catch a half-written agent binary.
- The agent is built after MAROC_STRIP_DEBUG_SYMBOL, which runs GNU strip
  across $(RK_APP_OUTPUT)/bin. That is not something to point at a Go binary,
  and -s -w already did the equivalent at link time.

Cross-compiling for 32-bit turned up a real bug: unix.Sysinfo_t's numeric
fields are uint64 on 64-bit hosts but uint32 on 32-bit, while statusResponse
declares uint64, so /status did not compile for the target at all. The agent
had only ever been built for amd64. Fixed with explicit conversions, which are
no-ops on 64-bit.

(Left alone: Totalram and Freeram are counts of sysInfo.Unit bytes, not bytes.
mem_unit is 1 on Linux and this device has 256MB, so nothing overflows, but
the field is reported as though it were bytes.)

S99motionsense gains the agent's lifecycle: started after the C daemon so its
log does not interleave with the camera handover, guarded by [ -x ] so a
missing agent does not stop MotionSense, and stopped first on shutdown.
Ordering is not actually load-bearing — the agent treats a missing database as
non-fatal and retries the socket — but it keeps the logs readable.

One new build requirement: agent/ has no vendor/ directory, so a cold build
downloads Go modules and needs network. The SDK build already needs network
for buildroot sources, so this is not a new kind of dependency, but it does
mean the build cannot run fully offline.

Verified by a full build: both binaries in /oem/usr/bin, oem 16.25/60MB.

+++ agent/internal/server/api.go
@@ -85,9 +85,12 @@ func statusHandler(broker *stream.Broker, start time.Time) func(http.ResponseWri
 		status.AppUptime = time.Since(start).String()
 		status.Clients = broker.ClientCount()
 		status.SystemUptime = (time.Duration(sysInfo.Uptime) * time.Second).String()
-		status.WorkLoad = sysInfo.Loads[0] / 65536.0
-		status.Totalram = sysInfo.Totalram
-		status.Freeram = sysInfo.Freeram
+		// unix.Sysinfo_t's numeric fields are uint64 on 64-bit but uint32 on
+		// 32-bit, which is what the RV1106 target is. Convert explicitly so
+		// this compiles for both.
+		status.WorkLoad = uint64(sysInfo.Loads[0]) / 65536
+		status.Totalram = uint64(sysInfo.Totalram)
+		status.Freeram = uint64(sysInfo.Freeram)
 
 		w.Header().Set("Content-Type", "application/json")
 		err := json.NewEncoder(w).Encode(status)
